# Speed up file system uploads